[Fix] ExtensionのIndicatorViewを廃止し、検索画面で直接実装する ExntesionではIndicatorViewが表示されないことがあったため
